Split out domForHtml
